requirements.txt
fastapi
uvicorn
python-multipart
pypdf
mistralai
numpy
streamlit
python-dotenv
pytest
reportlab
